+++ R/felt_layers.R
@@ -18,8 +18,6 @@ create_felt_layer <- function(map_id,
                               layer,
                               name = NULL,
                               token = NULL) {
-  map_url <- felt_map_url_build(map_id)
-
   check_string(layer, allow_empty = FALSE)
 
   if (is_url(layer)) {
@@ -51,6 +49,8 @@ create_felt_layer <- function(map_id,
       )
     )
   }
+
+  map_url <- felt_map_url_build(map_id)
 
   cli_alert_success(
     "Layer {.val {data$attributes$name}} created at {.url {map_url}}"
@@ -122,7 +122,8 @@ delete_felt_layer <- function(map_id,
 #' @export
 read_felt_layers <- function(map_id,
                              simplifyVector = TRUE,
-                             token = NULL) {
+                             token = NULL,
+                             call = caller_env()) {
   resp <- request_felt(
     endpoint = "read layers",
     map_id = map_id,
@@ -145,4 +146,68 @@ read_felt_layers <- function(map_id,
   }
 
   layers
+}
+
+
+#' Get Felt layer styles or update layer styles
+#'
+#' @description
+#' `r lifecycle::badge("experimental")`
+#'
+#' Get a Felt layer style or update a layer style. Warning, updating a layer
+#' style without a list that can be converted to a valid Felt Style Language
+#' (FSL) may get a layer into an *irreversible broken state*.
+#'
+#' @inheritParams request_felt
+#' @param layer_id If `NULL` (default), all layers for the map are used.
+#'   Multi-layer maps are not currently supported. Otherwise use a layer ID
+#'   string. Use [read_felt_layers()] to list layers for an existing map.
+#' @param style A named list that can be converted to a valid Felt Style
+#'   Language string. If style is supplied with a datasets id value matching the
+#'   layer datasets ids, this function updates an existing layer style. If style
+#'   is `NULL` (default), read styles for supplied map and layer. See the
+#'   [documentation on the Felt Style
+#'   Language](https://feltmaps.notion.site/Felt-Style-Language-de08179cb8494d3d8bbf5fb970f03fd0)
+#'   and the [API endpoint for updating layer
+#'   styles](https://feltmaps.notion.site/Felt-Public-API-reference-c01e0e6b0d954a678c608131b894e8e1#722105ec74e7492cb934bf81338db8b5)
+#'   for more information.
+#' @rdname felt_layer_styles
+#' @export
+felt_layer_styles <- function(map_id,
+                              layer_id = NULL,
+                              style = NULL,
+                              call = caller_env()) {
+  layer_id <- layer_id %||%
+    read_felt_layers(map_id, call = call)[["id"]]
+  check_string(layer_id, allow_empty = FALSE, call = call)
+
+  endpoint <- "get layer style"
+  data <- NULL
+
+  if (!is_null(style)) {
+    endpoint <- "update layer style"
+
+    cli_ifnot(
+      is_list(style) && all(has_name(style, c("datasets", "visualizations"))),
+      "{.arg style} must be a list with names {.val datasets}
+      and {.val visualizations}",
+      .fn = cli::cli_abort,
+      call = call
+    )
+
+    data <- list("style" = style)
+  }
+
+  # FIXME: Create a vectorized version of this for multiple layer_id values
+  resp <- request_felt(
+    endpoint = endpoint,
+    map_id = map_id,
+    layer_id = layer_id,
+    data = data
+  )
+
+  style <- httr2::resp_body_json(resp)[["data"]]
+
+  check_installed("RcppSimdJson")
+  RcppSimdJson::fparse(style)
 }