+++ NEWS.md
@@ -4,6 +4,7 @@
 * Add support for basemap and layer_urls parameter to `create_felt_map()` function.
 * Remove url parameter and add support for passing URLs as map_id parameter. (2023-07-13)
 * Add `get_felt_comments()` function (2023-07-13).
+* Add `felt_layer_styles()` function (2023-07-13).
 
 # feltr 0.1.1 (2023-06-30)